refactor: uint32 index

+++ src/pools/MultiShareClass.sol
@@ -82,7 +82,7 @@ contract MultiShareClass is Auth, IMultiShareClass {
     IPoolRegistry public poolRegistry;
     mapping(bytes32 salt => bool) public salts;
     mapping(PoolId poolId => uint32) public epochId;
-    mapping(PoolId poolId => uint8) public shareClassCount;
+    mapping(PoolId poolId => uint32) public shareClassCount;
     mapping(ShareClassId scId => ShareClassMetrics) public metrics;
     mapping(ShareClassId scId => ShareClassMetadata) public metadata;
     mapping(PoolId poolId => mapping(ShareClassId => bool)) public shareClassIds;
@@ -110,7 +110,7 @@ contract MultiShareClass is Auth, IMultiShareClass {
     function addShareClass(PoolId poolId, string calldata name, string calldata symbol, bytes32 salt, bytes calldata) external auth returns (ShareClassId shareClassId_) {
         shareClassId_ = previewNextShareClassId(poolId);
         
-        uint8 index = ++shareClassCount[poolId];
+        uint32 index = ++shareClassCount[poolId];
         shareClassIds[poolId][shareClassId_] = true;
 
         // Initialize epoch with 1 iff first class was added
@@ -543,7 +543,7 @@ contract MultiShareClass is Auth, IMultiShareClass {
     }
 
     /// @inheritdoc IShareClassManager
-    function previewShareClassId(PoolId poolId, uint8 index) public pure returns (ShareClassId scId) {
+    function previewShareClassId(PoolId poolId, uint32 index) public pure returns (ShareClassId scId) {
         return newShareClassId(poolId, index);
     }
